Give a name to prepended module
Currently, when you use `prepend` on an invisible module, you get a
dup'ed copy of the module in the class' ancestors, because Invisible has
to copy the original module and change method visibility to match the
prepending class.

This change gives a name to this dup'ed mod so it's easier to recognize
in the class' ancestors.