Only prepend dup'ed module once if possible
Normally, you can prepend a module many times and it will only actually
be prepended once. But if you `extend Invisible`, your module will now
not obey this rule.

This change fixes this *given the module has a name*. If it has no name,
this would be more complicated to do, so I'm just going to make that
assumption, which holds true for most cases.